Addressed comments and cleaned up. - changed processor function to annotate IDRs first before applying links in martinize. - moved cross go potential removal to water_bias.py - changed aminoacids.ff to reflect new naming - removed no longer needed import in dssp/__init__.py - changed name of import in processors/__init__.py
